+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # CHANGELOG
 
+## Version 0.13 - 02/04/2025
+
+### Added
+* Added `--output-type`, `--content-type` arguments to SPG Server.
+* Added `argument_seperator` const string to spg_lib.d
+* Added `loadfile` command.
+* Added `{code}` and `{code_end}` magics.
+* Added `{scrool}` and `{scroll_end}` magics.
+
+### Changed
+* SPG Server now prints output type of resource to stdout.
+* `__version__` variable changed to `0.13` in spg_lib.d
+* `check_exact_arg_count` function added to spg_lib.d
+* Informations on DOCUMENTATION.md file.
+* Argument parsing is changed, from now on commands can have ':' in their arguments. If a command is requies multiple arguments you can use ',,' to seperate arguments.
+* `examples/showcase.spg` is modified for better.
+
+### Removed
+* Unused `Error` struct is removed from spg_lib.d
+
 ## Version 0.12 - 24/03/2025
 
 ### Added